Add files via upload
INTRODUCTION:
The increased usage of private vehicles in metropolitan areas has resulted in significant rise in fuel consumption’s that have adverse effect on the climate.
This has led people in today’s society to accept problems like road traffic as the norm
In that context, the Bike Share initiative was launched to tackle the public mobility problem.
In a bike-share system, a user could lend a bike from any bike stations and return it to a bike station near the destination.
By using machine learning models ,we have predicted the bike sharing demand.
PROBLEM STATEMENT:
The crucial part is the prediction of bike count required at each hour for the stable supply of rental bikes, which can help people in predicting the bike demands. 
OBJECTIVES:
To develop machine learning models to predict the number of bike required in a Bike Sharing System. 
Bike-sharing systems to allow any one to hire bicycles from one of the city's numerous automated rental stations, ride it for a short distance, and then return it to any station in the city.
The main objective is to make predictive model, which could help them in predicting the bike demands proactively. 
To help them in stable supply of bike wherever needed.
CONCLUSION:
 Linear regression ,decision tree & random forest algorithms are used for model building.
Using the proposed models which provides a reasonable benchmark for evaluating the efficacy of the models.
The primary objective of the Internship was to gather a real time working experience and put the theoretical knowledge into practice.